revert revapi failures and move classes around
