+++ core/src/main/java/org/apache/iceberg/TrackedFile.java
@@ -26,7 +26,7 @@ import org.apache.iceberg.types.Type;
 import org.apache.iceberg.types.Types;
 
 /** A file tracked by a manifest. */
-public interface TrackedFile {
+interface TrackedFile {
   Types.NestedField TRACKING =
       Types.NestedField.required(
           147, "tracking", Tracking.schema(), "Tracking information for this entry");
