+++ core/src/test/java/org/apache/iceberg/TestInclusiveStatsEvaluatorWithTransforms.java
@@ -23,7 +23,6 @@ import static org.apache.iceberg.StatsTestUtil.fieldStats;
 import static org.apache.iceberg.StatsTestUtil.trackedFile;
 
 import org.apache.iceberg.expressions.Expression;
-import org.apache.iceberg.expressions.InclusiveStatsEvaluator;
 import org.apache.iceberg.expressions.TestInclusiveMetricsEvaluatorWithTransforms;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 import org.apache.iceberg.types.Types;
